Specification is supported slightly

+++ src/test/resources/psi/ProgramWithoutName.f
@@ -1,3 +1,3 @@
-integer a
+integer(kind=4) a
 integer b, c
 end program
